Indent theme.json with spaces instead of tabs
I'm going to turn this into YAML, then generate from the YAML back to
this. generate.js uses spaces so git diff sees changes in every line.

But I want to compare if the values changed, not the whitespace.

Also removes a trailing comma that breaks the JSON to YAML conversion.