Use the YAML architecture from the Dracula theme
YAML is a much better choice for configurations than JSON is. I can use
variables (like a centralized color palette I can reference for every
tokens) and comments. :tada: Plus boilerplate is kept to a bare minimum.

On the other hand, this means going against convention and I lose access
to useful helpers like hovers, color decorators and pickers, etc.

But the peeps at the Dracula theme for VSCode [1] have developed a
pretty cool set of scripts to support debugging of themes (no hot-reload
though) and linting (from the terminal, not integrated with vscode) and
they make YAML arch workable enough.

Because I'll only be copying Lucario/Atom, without making color scheme
choices, I don't think I'll have much of a need for color decorators and
the like. Furthermore, those will still be available in the generated
JSON theme, which I can edit as if testing styles in a browser's
DevTools, then propagate the changes to the source after I'm happy with
the final result.

Anyways, this commit adapts the script and the setup in the
dracula/vscode project so they suit my needs. I've yet to pull the
lint script, cause I don't think I'll need that for the time being.

Also, I made the initial JSON to YAML conversion with [2], and
regenerated it (no changes) with the generate + build scripts.

[1] https://github.com/dracula/visual-studio-code
[2] https://json2yaml.com/