Copy Lucario/Atom syntax theming rules
The ones in [1]. But I anticipate I'll rollback a great deal of them
because it makes langs like Ruby or Rust impracticable, their code a
colorful fruit salad where white/fg is the exception rather than the
norm.

I think it's because it sets variables (instance, library, any kind) to
Orange (which is more of a Red but whatever).

[1] https://github.com/raphamorim/lucario/blob/master/styles/language.less