Rework git gutter indicators
I dropped the idea of using alpha-transparent versions of the base
colors for this because they are like two very different things (red,
yellow, green for syntax vs for git status). I don't see a need for them
to be similar.

I dropped the idea of using blue/purple for modified because
green-yellow-red is an universal code (except for Japan, I read XD).

I'm keeping the atom.Yellow code as ModYellow, and Sublime's DiffGreen
and DiffRed as AddGreen and DelMagenta.

Atom's Yellow is familiar to me, whereas I don't particularly fancy its
Red (too similar to the syntax one), nor its White/FG for Added (wtf?).
I'm fine with Sublime's Green and Magenta, the later shared with
Lucario/Vim too.

As a matter of fact, I found out that Sublime's Green and Magenta come
from Monokai: its Green and Pink respectively. And the base scheme's
Yellow and Cyan are Monokai Yellow and Blue.