Improve line number colors
This commit makes them less prominent (they were ~= foreground).

I tried to reproduce the color in Atom (#a6aeb0) and got close enough
(#a4acb0) with foreground + alpha 9c. But this leaves too little margin
for having the active line number be significantly more visible.

So I tried alpha 64, and it's a pretty noticeable difference (vs
foreground) but then I thought we don't need the active lnum being so
prominent, at the expense of other lnums being less clear. Because this
is a theme that uses a pretty visible current line highlight.

Unlike e.g.: Night Owl, which has a negligible Line Highlight which it
compensates with painting the active lnum even brighter than default FG.

So I settled for 'FG alpha 80' vs 'FG', and it looks just fine.

Also, I made the folding control (arrow next to lnums that folds
ranges / scopes of code) the same color as active lnum, which seemed to
be a convention in the default styles for them.