Color surrounding frame/bars the Dracula way
That is, by using shades of the main background color. Though the detail
of coloring the sideBar border is a personal touch.

Well, I saw it somewhere else I can't remember. It fakes a subtle shadow
between the sidebar and the editor.

The color for BgDarker is from Atom's One Dark UI theme when it
interacts with the Lucario theme. That is, it's a shade of Lucario's
canonical BG (very similar HS), set to whatever Value/Lightness drop
One Dark UI is programmed to work with.