Theme Markdown links
Terminology: [link title](link url)

Preserves the color for link titles, StringYellow, which were already
themed because of a blanket "string" scope match.

But now we also colorize the brackets around link titles with the same
yellow color, in the spirit of other inline punctuation marks like the
asterisks in *italics* or **bold**.

For link urls I considered EntityRed, but OrangeInstance makes for a
better collocation with StringYellow.

I don't like how urls are underlined though, it's too attention-grabbing,
and the important (readable) part is the title. But that's on vscode,
cannot change it from a theme.