YAML: swap anchor and alias color
It makes more sense to me that anchors are white, and aliases are
orange. It makes more sense to me, at least on cmd-help.sublime-syntax.

Considering that our orange and yellow are pretty close, that anchors
usually appear next to strings (yellow), that aliases substitute strings
(often more than any other scalar) in the value side of YAML keys.

Don't be surprised if I revert this. :sweat_smile: