md: fix specificity of a couple of scope selectors
Detected while programming in python, that list begins "[" are colored
in pink (unlike list ends "]"). This commit fixes that.

Also increases the specificity of the code fences scope selector, to
circumscribe it to Markdown only. It's unlikely to pop up in any other
lang, I guess, but just to be sure.

Every other Markdown selector includes the ".markdown" suffix or is
scoped with "text.html.markdown" before the particular scope.