Add tasks.json configuration
So launching the Extension Host doesn't unnecessarily reveal the
terminal panel. Now it'll only open the panel (on the Problems tab) if
the build fails.

This commit doesn't make any changes to the Extension Host's (pre) build
step, if I understood correctly. It was doing the exact same thing
before, with an implicit tasks.json config identical to this new one.
With the sole exception of not revealing the terminal panel on success.