Add workspace setting set by Quick Theme Switcher
I'm using this extension [1] to quickly compare with other themes (in
particular: Dracula and Night Owl; and Chinolor Light later for the
syntax support).

This extension changes the theme in workspace settings (I'd rather in
user's but whatever) so I'm tracking this in order to keep it away from
git's dirty status. As much as possible, cause it'll be a different
theme quite often.

I intend to remove this in due time: when I roll out my own extension
for this, or/and when 'deadline mode' is behind me.

[1] sandorex.vscode-theme-quick-switch