Colorize the Extension Host's status bar in *Accent
To better differentiate this window from the actual project one.
Especially because now it features as a fully fledged project/workspace
in "Open Recent".

Do so by leveraging the Peacock extension:
https://marketplace.visualstudio.com/items?itemName=johnpapa.vscode-peacock

This commit adds the config for that extension, and so it only colors
the status bar and not any other part of the UI. The status bar is
pretty stable. Other bars, not so much.