Fix "task X cannot be tracked" error
Wasn't breaking, but a hassle all the same. Plus adds a 5s timeout.

Explanation: the Debug Host launcher needs to know that the preLaunch
task (npm: watch) has finished before launching.

So an "endsPattern" that should match the task's output is used to
determine if it has finished. Or, in our case, if it's ready.

Still doesn't close the watcher on Extension Host exit though. Shitty.

Useful reference:
https://github.com/microsoft/vscode/issues/114933