Add postDebugTask to stop npm:watch
VSCode doesn't have a straightforward method to kill watch tasks from
another task or launcher, so I turned to the system. I leverage pgrep so
it'll probably not work on Windows...

VSCode's Terminate Task command was promising. There are ways to use
commands as tasks [1-3]. But it lacks support for passing the task to
kill as an argument.

Solutions using Terminate Task either kill all tasks or rely on the
relevant task being sorted first. And the later didn't even work for me.

Task label would work, it's already used as id in tasks and launch
configs (dependsOn, preLaunchTask, etc.).

In any case, pgrep works and I'm happy with this.

[1] https://stackoverflow.com/questions/54480792/terminate-another-task-from-within-a-postdebugtask-vs-code
[2] https://github.com/microsoft/vscode/issues/90786
[3] https://code.visualstudio.com/docs/editor/variables-reference