Overhaul UI colors to resemble Atom OneDark UI
So yeah, this is the true essence of Lucario for Atom that I used to
love.

I became hung up on that little effect of border shadows between side
bar and editor, but they're not worth ditching OneDark UI.

This is a first step only, there's a lot of work to do in this direction
and I have to tweak everything. Even the UI colors are subject to
change.