Set color for git ignored files
Git colors are future work (UI red, yellow, green are TBD yet) but I
strongly dislike how the default color for this looks tan against the
theme's blue-hued backgrounds.

fg-subtle is also considerably darker (or alpha-ed). Which is also good
because the default color was almost as bright as fg-base.