Highlight the not-current find matches with border
As opposed to background color. Atom One Dark UI does it like this. It's
nice, imo.

Though Atom makes the corners on those borders rounded, which looks a
bit better than the harsh squares here.

Plus I hate how the default VS Code Dark color for this is looks
brown-ish on top of Lucario's blue bg.

This is a test, not final by any means.