Increase the alpha of the current find match
The default one has a lower alpha. It's a bit easier to locate the
current search match like this.

On the other hand, it makes the matched text look totally overshadowed
when it's on a comment. (Comments use a _pastel_ blue color.) For every
other syntax color, the text is visible enough.

Besides, the last character of the current match gets a black foreground
when vscodevim is enabled. For whatever reason. I hoped to find a fix
for that with this foray... Shitty.