Make the menu separator color darker
I added a new anchor, fg-separator, which is an alpha of fg-base, with
much more transparency than fg-subtle.

The cool part about making it alpha-based is that, I hope, it will work
well as a separator for all shades of background (bg-level-1 to 3).

I'm thinking of doing the same for border-base, expressing it as a pure
black with an alpha channel...