Add yaml-color-hex schema to allow YAML features
The schema for color-theme.json doesn't allow a couple of features that
we're able to use in our YAML themes, namely:

- nulls: I left undefined keys all over the theme, either as placeholders
  (colors to be defined) or with "been there, won't define" comments.
- the !alpha YAML tag that combines a color with an alpha channel.

We're able to do this because we have a YAML-to-JSON compile step
(generate.js) were we cull keys without value (null) and we transform
!alpha tags into proper hex colors.

This new schema covers those 2 cases as a union type with the existing
type:string + format:color-hex schema.

format:color-hex seems to be a custom vscode addition btw, couldn't find
it documented or discussed anywhere.

The 3rd option (arrays) is to allow !alpha + sequence. But it's a very
weak validation, allowing all kind of incorrect values as long as
they're a sequence.

But I don't know if there's a way to validate that they're prefixed by
"!alpha". Also, validation is the lesser of my concerns for enabling the
json schema, the killer feature are the hover tooltips for workbench
properties.

The changes to the theme and to workbench-colors.json are my testing,
included here to demonstrate the change. Now those properties I've
mapped to yaml-color-hex don't throw errors.