Restrict the !alpha schema validation a little more
With this, only arrays of length 2 will be allowed.

I tried the prefixItems validator to restrict the 2 items to
string+color-hex & string of length 2. But it doesn't seem to work? Not
in vscode-yaml at least.

In any case, this + the workspace setting (yaml.customTags) that only
allowlists !alpha, should be good enough.

Reference:
https://json-schema.org/understanding-json-schema/reference/array