Refactor VERSION_NUMBER variable to be an env
I had a previous version of all this that used the VERSION_NUMBER
variable in the action-gh-release step.

We don't have access to (Bash) env variables [1] in an action's input
parameters, so I had to use a step output parameter [2]. The latter
don't have that restriction.

But then I refactored it all into the 'Gather release assets' step and
now we can use plain envs, which are much more sucint.

I'm keeping the output parameter just in case. e.g.: if I use an action
to publish to the VS Marketplace, I'll probably need it to point to the
VSIX package.

[1] https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-environment-variable
[2] https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-output-parameter