Explicitly link GitHub releases to a tag
On the first versions of the Action, which I had copied from the
Material Icon Theme extension, tag_name was explicitly set in the
sub-action that releases to GitHub. [^1]

Before committing the first version of this, I did a bit of a clean-up,
including deleting the tag_name from that action's inputs. It shouldn't
be needed because it defaults to github.ref [^2]. github.ref is set to
the pushed tag that triggers the whole Action.

After that, I changed the vTEST tag to v0.0.0 because vsce wouldn't
package for a version not complying with semantic versioning.

Then I worked on uploading the JSON and VSIX assets to the release. But
I noticed that the .zip and .tar.gz compressions of the repo were no
longer listed in releases.

So I debugged. It turns out that the release isn't even linking to a
tag. That's because when testing with Act, we're not really in the
context of a GitHub push, and so `github.ref` is unset.

This commit addresses restores tag_name and so we can test in peace.
It'll be redundant when we execute on GitHub but we can live with that.
Redundancy over magic.

[^1]: https://github.com/PKief/vscode-material-icon-theme/blob/main/.github/workflows/release.yml#L64
[^2]: https://github.com/softprops/action-gh-release?tab=readme-ov-file#inputs