delivery: Extract Changelog -> Release Body logic into a script
Simplifies the workflow file, enables syntax highlighting on the script,
etc. Makes more sense like this.

I considered collocating this with the workflow configuration, but it
may be useful to use elsewhere and it's not like I don't already have
utility scripts in scripts/, with adapt-schemas.js

Also: stub the Changelog version for testing purposes.