delivery: Pin action version to the commit SHA
Only for action-gh-release, the other 3rd-party actions are owned by
GitHub (action/ namespace).

GHA are susceptible to supply chain attacks if you refer to them by the
version (a git tag, which can be easily re-addressed). The advisory is
to use the commit SHA for that particular version instead.

References:

https://docs.github.com/en/actions/security-for-github-actions/security-guides/security-hardening-for-github-actions#using-third-party-actions
https://semgrep.dev/blog/2025/popular-github-action-tj-actionschanged-files-is-compromised/
https://alexwlchan.net/2025/github-actions-audit/