patch-schemas: remove a TODO
It doesn't make sense to make the replacement condition more strict
because JSON Schema objects for e.g., colors:editor.background, also
include keys like description, default or defaultSnippets.

Also, in the process of trying to do this I've become painfully aware
that this project really is not the proper place to handle json schemas,
they belong in a separate project.

- I went back to v1.47.3 for the schemas, to test changes to the script,
- but then schemas/*.json changes were hidden in the git diff,
- then I had to comment out all of the .gitattributes code that I had
  just added in this branch...

DX is so shitty that I really am tempted to refactor this out into its
own project.